Separate service handling code from main code
This patch adds a Serv type to encapsulate the data and handler for an
individual service. Handling code specific to NickServ and ChanServ is
separated out of the Arata module into Arata.NickServ and
Arata.ChanServ, making use of the Serv type to expose their data and
handlers to Arata.

Per-service hosts are removed as they can cause inconsistencies and
problems, the most notable of which is the inability to send a PRIVMSG
directly to `nick@host` due to the `@host` syntax requiring the actual
remote server name. This patch amends the client introduction code in
the Arata module to set service hosts to the existing server name config
option in the info section of the config file.