NickServ: Move HELP handling to new module
This patch separates out NickServ HELP handling from the main NickServ
module and lays the foundation for a proper plugin API with the ability
to expose new services and to expose new commands to specific services.

The new Help module is in a temporary state but demonstrates proof of
concept code to display help text by looping over a map of commands.