Add Dockerfile
This allows people to easily run Arata inside docker for local
evaluation or for people who don't want to install the haskell
dependencies on their system. This also makes deployment easier (sans
the database, but that can be worked on in the future).

Amendment 1: Discover the ircd to link to

Amendment 2: fix an if logic statement