Protocol.Charybdis: Fix pattern match failure
`protoHandleMessage` in the Charybdis protocol module previously made an
irrefutable pattern assertion about the result of attempting to retrieve
a client from the environment state matching the UID prefix of the ENCAP
command.

This patch makes use of LambdaCase to pattern match on the result of
`getClient` for both `Nothing` and `Just src`. In the case of `Nothing`,
no action is performed. In the case of `Just src`, the existing action
to handle the ENCAP message is performed.

Fixes #4.