Redo protocol system entirely
This patch replaces the old hardcoded protocol system with the ability
for plugins to export functions for Message -> [Event] and Action ->
[Message], as well as a function to convert from an Integer UID to a
String UID.

The protocol module now only performs these functions and any handling
of the events returned for a given Message is done in the core of Arata.
Similarly, command handlers now have a result type of [Action] and additional
handling of each action can be done in the core of Arata.

This required a very large chunk of the code base to be rewritten,
including every plugin exporting a command handler. However, this design
decision is for the greater good as it ensures that all events and
actions are always handled consistently and it will also go a long way
towards preventing the code base from becoming unmanageable due to
widespread duplicate code and inconsistencies.

Closes #12.