Index and lookup account nicks/names by lowercase
This patch amends the Indexable instance for Account to `map toLower`
over the account name. The DB helper functions `getAccountsByNick` and
`getAccountsByName` are also amended to `map toLower` over the input
nick/name before doing a lookup on the index.

Use of the `toLower` function should be replaced with an RFC 1459
compliant case mapper for case insensitivity to hold true for account
names containing the symbols `{}|` which map respectively to `[]\`

Addresses #13.